+++ src/main/java/com/ajou_nice/with_pet/pay/model/dto/PayReadyResponse.java
@@ -1,4 +1,4 @@
-package com.ajou_nice.with_pet.domain.dto.kakaopay;
+package com.ajou_nice.with_pet.pay.model.dto;
 
 
 import java.time.LocalDateTime;
